// src/app/rick-and-morty-wiki/interfaces/locations.interface.ts
export interface APIDataLocations {
  info:    Info;
  results: Location[];
}

export interface Info {
  count: number;
  pages: number;
  next:  string;
  prev:  string;
}

export interface Location {
  id:        number;
  name:      string;
  type:      string;
  dimension: string;
  residents: string[];
  url:       string;
  created:   Date;
}